# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meridian API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/approvals/{loan_id}/{gate}/decision": {
      "post": {
        "tags": [
          "approvals"
        ],
        "summary": "Post Approval Decision",
        "operationId": "post_approval_decision_api_v1_approvals__loan_id___gate__decision_post",
        "parameters": [
          {
            "name": "loan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loan Id"
            }
          },
          {
            "name": "gate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Gate"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
    "/api/v1/evaluations": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Start Evaluation",
        "operationId": "start_evaluation_api_v1_evaluations_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartEvaluationResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Get Evaluation",
        "operationId": "get_evaluation_api_v1_evaluations__evaluation_id__get",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationStatusResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{slug}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Document",
        "operationId": "get_document_api_v1_documents__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "architecture",
                "assumptions"
              ],
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApprovalDecisionBody": {
        "properties": {
          "artifact_digest": {
            "type": "string",
            "title": "Artifact Digest"
          },
          "decision": {
            "type": "string",
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Decision"
          },
          "approver": {
            "type": "string",
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Approver"
          }
        },
        "type": "object",
        "required": [
          "artifact_digest",
          "decision",
          "approver"
        ],
        "title": "ApprovalDecisionBody",
        "description": "`loan_id` and `gate` are path parameters; the body carries only the decision\ninputs plus the artifact digest the browser must echo back unchanged."
      },
...
      "DocumentResponse": {
        "properties": {
          "slug": {
            "type": "string",
            "enum": [
              "architecture",
              "assumptions"
            ],
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "markdown": {
            "type": "string",
            "title": "Markdown"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "title",
          "markdown"
        ],
        "title": "DocumentResponse"
      },
      "ErrorEnvelope": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorEnvelope"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "active_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Run Id"
          },
          "active_evaluation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Evaluation Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
      "EvaluationReportResponse": {
        "properties": {
          "exists": {
            "type": "boolean",
            "title": "Exists"
          },
          "markdown": {
            "type": "string",
            "title": "Markdown"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "modified_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At"
          }
        },
        "type": "object",
        "required": [
          "exists",
          "markdown",
          "path"
        ],
        "title": "EvaluationReportResponse"
      },
      "EvaluationStatusResponse": {
        "properties": {
          "evaluation_id": {
            "type": "string",
            "title": "Evaluation Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "returncode": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returncode"
          },
          "output": {
            "type": "string",
            "title": "Output",
            "default": ""
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationReportResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "terminal_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal At"
          }
        },
        "type": "object",
        "required": [
          "evaluation_id",
          "status",
          "started_at"
        ],
        "title": "EvaluationStatusResponse",
        "description": "A reconnectable snapshot consumed by the evaluation page."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StartEvaluationResponse": {
        "properties": {
          "evaluation_id": {
            "type": "string",
            "title": "Evaluation Id"
          },
          "status": {
            "type": "string",
            "const": "running",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "evaluation_id",
          "status"
        ],
        "title": "StartEvaluationResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}